requirements.txt
streamlit
requests
beautifulsoup4
Pillow
opencv-python
rembg
